Two Changes: No longer format solarize at cursor & stop autocomplete once regex type is found in line
No longer format solarize at cursor: A bug was generated by using G.quill.format to a color span to color the text at the cursor spot.  Caused jumping on the line, especially if using the arrow key.  Fixed.

Stop autocomplete once regex type is found: Once we find an autocomplete of a specific type, we no longer try to autocomplete that type in the sentence.  For example, once an operator is found, we no longer try to autocomplete operators in that line.